requirements.txt
streamlit
chromadb
sentence-transformers
google-generativeai
pandas
numpy
uuid
argparse
python-dotenv

langchain
langchain-google-genai
google-genai